+++ apps/frontend/src/components/ConfigPluginDialog.tsx
@@ -9,7 +9,7 @@ import Add from '@mui/icons-material/Add'; // For AddButton
 import BluetoothIcon from '@mui/icons-material/Bluetooth'; // For selectDevice icon=ble
 import DeleteForever from '@mui/icons-material/DeleteForever'; // For RemoveButton
 import DeviceHubIcon from '@mui/icons-material/DeviceHub'; // For entities icon=matter
-import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline'; // For ErrorListTemplate
+import ErrorOutlineIcon from '@mui/icons-material/ErrorOutlineOutlined'; // For ErrorListTemplate
 import HubIcon from '@mui/icons-material/Hub'; // For selectDevice icon=hub
 import KeyboardDoubleArrowDownIcon from '@mui/icons-material/KeyboardDoubleArrowDown';
 import KeyboardDoubleArrowUpIcon from '@mui/icons-material/KeyboardDoubleArrowUp';
@@ -1220,7 +1220,7 @@ export const ConfigPluginDialog = ({ open, onClose, plugin }: ConfigPluginDialog
   if (!open || !schema || !formData) return null;
   return (
     <Dialog open={open} onClose={onClose} slotProps={{ paper: { sx: { maxWidth: '800px' } } }}>
-      <DialogTitle gap={'20px'}>
+      <DialogTitle>
         <div style={{ display: 'flex', flexDirection: 'row', alignItems: 'center', gap: '20px' }}>
           <img src="matterbridge.svg" alt="Matterbridge Logo" style={{ height: '32px', width: '32px' }} />
           <h3>Matterbridge plugin configuration</h3>