+++ apps/frontend/src/components/Header.tsx
@@ -1,7 +1,7 @@
 // @mui/icons-material
 import DownloadIcon from '@mui/icons-material/Download';
 import Favorite from '@mui/icons-material/Favorite';
-import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
+import HelpOutlineIcon from '@mui/icons-material/HelpOutlineOutlined';
 import HistoryOutlinedIcon from '@mui/icons-material/HistoryOutlined';
 import HomeOutlinedIcon from '@mui/icons-material/HomeOutlined';
 import MoreHoriz from '@mui/icons-material/MoreHoriz';